+++ docs/docs/news.md
@@ -13,6 +13,7 @@ sidebar_label: Changelog
 A new family of `update_*` functions to easily update graph aspects on the fly.
 
 - `update_node_size` and `update_node_color` to update a single node.
+- `update_nodes_size` and `update_nodes_color` to update multiple nodes from a data.frame.
 - `update_link_source_color` and `update_link_target_color` to update a link source and target color.
 
 # 0.1.0